requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pdf2image>=1.16.0
cairosvg>=2.7.0
pypdf>=3.0.0
google-generativeai>=0.3.0
